Removes code that causes a javascript error
Additionally, the piece of code doesn't do what it says.
The comment indicates that it removes only empty attributes, but
while it actually removes all class title and target attributes.

The reason why this is done is not specified either, so I don't
see why this code exists at all.